fix: arreglo de manejo de sesión y botones (incluyendo NavBar) relacionados
